requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
blinker==1.6.2
bokeh==3.1.1
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.0
Flask==2.3.2
flatbuffers==23.5.8
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
ipywidgets==8.0.6
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter-bokeh==3.0.7
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.6
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
panel==0.14.4
param==1.13.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyct==0.5.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyviz-comms==2.2.1
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.4
widgetsnbextension==4.0.7
wordcloud==1.9.1.1
wrapt==1.14.1
xyzservices==2023.2.0
yarl==1.9.2
